:sparkles: context provider 위치 수정 및 useRouter로 페이지 이동 구현
