--- src/pages/WriteLetter/index.tsx
@@ -1,12 +1,0 @@
-import { LetterContextProvider } from './hooks/useLetterContext';
-import CreateLetter from './CreateLetter/CreateLetter';
-
-export default function WriteLetter() {
-  return (
-    <>
-      <LetterContextProvider>
-        <CreateLetter />
-      </LetterContextProvider>
-    </>
-  );
-}
