+++ src/pages/letter/hooks/useLetterContext.tsx
@@ -5,8 +5,6 @@ interface LetterContextProps {
   setLetterImage: React.Dispatch<SetStateAction<File | null>>;
   letterText: string;
   setLetterText: React.Dispatch<React.SetStateAction<string>>;
-  step: number;
-  setStep: React.Dispatch<React.SetStateAction<number>>;
 }
 
 const initialLetterContext: LetterContextProps = {
@@ -14,8 +12,6 @@ const initialLetterContext: LetterContextProps = {
   setLetterImage: () => {},
   letterText: '',
   setLetterText: () => {},
-  step: 1,
-  setStep: () => {},
 };
 
 export const LetterContext = createContext<LetterContextProps>(initialLetterContext);
@@ -27,15 +23,12 @@ interface LetterContextProviderProps {
 export const LetterContextProvider: React.FC<LetterContextProviderProps> = ({ children }) => {
   const [letterImage, setLetterImage] = useState<File | null>(null);
   const [letterText, setLetterText] = useState<string>('');
-  const [step, setStep] = useState<number>(1);
 
   const value: LetterContextProps = {
     letterImage,
     setLetterImage,
     letterText,
     setLetterText,
-    step,
-    setStep,
   };
 
   return <LetterContext.Provider value={value}>{children}</LetterContext.Provider>;
