+++ src/pages/letter/index.tsx
@@ -1,0 +1,9 @@
+import CreateLetter from './create-letter';
+
+export default function Letter() {
+  return (
+    <>
+      <CreateLetter />
+    </>
+  );
+}
